Extract main logic into distinct functions
